[no ci] Create new timetable file for rewrite
