+++ lib/main.dart
@@ -5,7 +5,7 @@ import 'package:eduapge2/homework.dart';
 import 'package:eduapge2/icanteen.dart';
 import 'package:eduapge2/load.dart';
 import 'package:eduapge2/messages.dart';
-import 'package:eduapge2/timetable.dart';
+import 'package:eduapge2/old_timetable.dart';
 import 'package:firebase_analytics/firebase_analytics.dart';
 import 'package:firebase_remote_config/firebase_remote_config.dart';
 import 'package:flutter/material.dart';
